update: tree cover loss max year
