+++ configs/dates-config.ts
@@ -1,0 +1,6 @@
+export const DATES = {
+  TREE_COVER_LOSS: {
+    min: 2001,
+    max: 2022,
+  },
+};
