+++ src/js/components/sharedComponents/TimeSlider.tsx
@@ -346,6 +346,7 @@ const TimeSlider = (props: TimeSliderProps): JSX.Element => {
     return label;
   };
 
+  console.log('this one is defnintely running', props);
   return (
     <div className="time-slider-container">
       {playButton ? (
